ref: remove injector
