+++ example/aws/parameter-store/main.py
@@ -2,7 +2,7 @@
 import hydra
 from omegaconf import DictConfig
 import logging
-from omegaconf_cloud_resolvers import CustomResolverInjector
+from omegaconf_cloud_resolvers import register_custom_resolvers
 from omegaconf_cloud_resolvers.resolvers.aws import AWSParameterStoreResolver
 
 logger = logging.getLogger(__name__)
@@ -16,7 +16,7 @@ resolvers = {
     ),
 }
 
-conf_injector = CustomResolverInjector.inject_resolvers(**resolvers)
+register_custom_resolvers(**resolvers)
 
 config_fname = "config"
 config_dir = "."
