--- example/aws/secrets-manager/config_with_defaults.yaml
@@ -1,2 +1,0 @@
-secret_1: "${aws_secretsmanager:test_secret_1}"
-secret_3: "${aws_secretsmanager:test_secret_2}"
