+++ example/gcp/secrets-manager/main.py
@@ -2,13 +2,14 @@
 import hydra
 from omegaconf import DictConfig
 
-from omegaconf_cloud_resolvers import CustomResolverInjector
+from omegaconf_cloud_resolvers import register_custom_resolvers
 from omegaconf_cloud_resolvers.resolvers.gcp import GCPSecretManagerResolver
 
 resolvers = {
     "get_gcp_secret": GCPSecretManagerResolver(),
 }
-conf_injector = CustomResolverInjector.inject_resolvers(**resolvers)
+
+register_custom_resolvers(**resolvers)
 
 config_fname = "config"
 config_dir = "."
