+++ omegaconf_cloud_resolvers/resolvers/aws/secretsmanager.py
@@ -1,6 +1,6 @@
 from .aux import try_cast_to_dict
 from .base import AWSMixin
-from ..base import Resolver
+from ..base import PluginResolver
 
 
 class AWSSecretManagerMixin(AWSMixin):
@@ -8,17 +8,17 @@ class AWSSecretManagerMixin(AWSMixin):
         return self._session.client("secretsmanager")
 
 
-class AWSSecretsManagerResolver(Resolver, AWSSecretManagerMixin):
-    def __init__(self, session=None, infere_json=False, *args, **kwargs):
+class AWSSecretsManagerResolver(PluginResolver, AWSSecretManagerMixin):
+    def __init__(self, session=None, infer_json=False, *args, **kwargs):
         super().__init__(session, *args, **kwargs)
-        self._infere_json = infere_json
+        self._infer_json = infer_json
 
     def __call__(self, name):
         secret = self.client.get_secret_value(SecretId=name)
 
         try:
             secret = secret["SecretString"]
-            if self._infere_json:
+            if self._infer_json:
                 return try_cast_to_dict(secret)
             else:
                 return secret
