+++ pyproject.toml
@@ -4,31 +4,28 @@ version = "0.0.0"
 description = "Omegaconf custom resolvers to retrieve configuration values from cloud services"
 authors = ["M Gil Valverde <m.gilvalverde@gmail.com>"]
 readme = "README.md"
-packages = [{include = "omegaconf_cloud_resolvers/"}]
+packages = [{ include = "omegaconf_cloud_resolvers/" }]
 
 [tool.poetry.dependencies]
 python = "^3.9"
 jmespath = "^1.0.1"
 omegaconf = "^2.3.0"
+hydra-core = { version = "^1.3.0", optional = true }
+boto3 = { version = "^1.34.16", optional = true }
+google-cloud-secret-manager = { version = "^2.20.2", optional = true }
 
-
-[tool.poetry.group.aws.dependencies]
-boto3 = "^1.34.16"
-
-[tool.poetry.group.gcp.dependencies]
-google-cloud-secret-manager = "^2.20.2"
-
-[tool.poetry.group.hydra.dependencies]
-hydra-core = "^1.3.0"
+[tool.poetry.extras]
+aws = ["boto3"]
+gcp = ["google-cloud-secret-manager"]
 
 [tool.poetry.group.dev.dependencies]
-moto = {extras = ["secretsmanager", "ssm"], version = "^5.0.13"}
+moto = { extras = ["secretsmanager", "ssm"], version = "^5.0.13" }
 pytest-mock = "^3.14.0"
 pytest = "^8.3.2"
 ruff = "^0.6.2"
 bandit = "^1.7.9"
 mkdocs = "^1.6.0"
-mkdocstrings = {extras = ["python"], version = "^0.25.2"}
+mkdocstrings = { extras = ["python"], version = "^0.25.2" }
 mkdocs-material = "^9.5.33"
 mike = "^2.1.3"
 python-semantic-release = "^9.8.7"
@@ -39,6 +36,6 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.mypy]
 disallow_untyped_defs = true
-exclude = ["_static", "build", "examples", "notebooks", "venv","_exp", "config", "dist", ".env", ".venv"]
+exclude = ["_static", "build", "examples", "notebooks", "venv", "_exp", "config", "dist", ".env", ".venv"]
